Set IsActive to false when received limit duration is 0
According to the EEBus UC Implementation Guideline for LPC v1.0.0
section 2.2 1):
If Actor Controllable System receives an Active Power Consumption Limit
with a duration set to "0", the limit SHALL immediately be deactivated
by the Controllable System (see [LPC-007]). If the received write
command contains a duration set to "0" AND also a request to set the
Active Power Consumption Limit to "activated", the Actor Controllable
System SHALL ignore the request

This has the added effect for us, that previously in the CS-LPC/CS-LPP
API it was next to impossible to differentiate between a limit without a
duration (active forever), and a limit with a 0 duration.